Adding an option for a client secret to use in a ClientSecretCredential (#164)
When using an application registration to authenticate, the usage of the
`InteractiveBrowserCredential` only works if this application
registration is used to authenticate a user, and that user is then used
to gain access to the ADT instance. But when using this application
registration as the actual authenticator (with _id_ and _secret_), we
need to use the `ClientSecretCredential` instead. To allow for this, a
new new `ClientSecret` configuration option is added which triggers the
usage of the `ClientSecretCredential`.