Add Mapping for ExternalID to Willow Ontology Mapping (#189)
### What
Adding a mapping to the externalId field in the Willow Ontology to
duplicate the TwinId as this supports backwards compatibility with older
apps at WIllow.

### Why
Some apps depend on having a "twinId" stored in the externalID field. 

### Tested
Ran locally and the externalID field is now populated in the Willow ADT
instance.