Allow for new relationship names (#207)
### What
Allow a formatter class to be passed in by the consumer to allow a
custom algorithm for setting the relationship Id

### Why
1. Willow has a different naming convention for relationship ids
2. It is possible to have multiple relationships between two entities of
the same type that differ by a property.
i.,e. a waterpump could feed both hot and cold water to a system, and we
might want to have a isFedBy for HotWater, and for cold water for the
twins.

### Tested
- Ran unit tests
- Integrated with the IntegrationManager.Mapped locally and ran those
tests. Will have a separate PR for those changes once this PR completes.