Handle Empty Twins properly (#212)
### What
Handle empty twins properly. This can happen if a twin is based on a
dtdl interface and the type cast requested doesn't match the type of the
object.

### Why
App was crashing in the situation where the twin id was empty because
the casting done to the interface did not match the actual type of the
instance of the twin

### Tested
Added test case to the site.json file. Verified previous code failed.
Added new code. Green tests.