Adding Support for Object Transformations (#219)
### What
Adding Support for Object Transformations

This is the first of multiple PRs for this change.

### Why
There are cases where there may be an object coming from the Source
Ontology that needs to be translated into a flat string property in the
target ontology (i.e. unit)

### Tested
Built and ran unit tests
Updated the packages that use this to build out the functionality needed
and built unit tests
Added all the packages to the application for topology ingestion and ran
that against a UAT environment and validated results.